+++ dsptoolbox/__init__.py
@@ -75,4 +75,4 @@ __all__ = [
     "effects",
 ]
 
-__version__ = "0.2.12"
+__version__ = "0.2.13"